+++ app/package.json
@@ -28,7 +28,7 @@
     "ethers": "^5.6.4",
     "expo": "~44.0.0",
     "expo-crypto": "~10.1.1",
-    "expo-dev-client": "~0.8.5",
+    "expo-dev-client": "^1.0.0",
     "expo-document-picker": "~10.1.3",
     "expo-image-picker": "~12.0.1",
     "expo-linking": "~3.0.0",
